Silence -Wsign-compare warning in sn76496
warning: comparison of integer expressions of different signedness:
'uint32_t' {aka 'unsigned int'} and 'int32_t' {aka 'int'}
… (m_RNG & m_whitenoise_tap2)!=(m_ncr_style_psg?m_whitenoise_tap2:0) …
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Honestly, it's hard to understand what was the author's intention with
this expression, but it's pretty clear the comparison of uint32_t vs
int32_t was unintentional. Left side turns into unsigned because m_RNG
is uint32_t.

+++ src/hardware/mame/sn76496.cpp
@@ -432,13 +432,10 @@ void sn76496_base_device::sound_stream_update(sound_stream &stream, stream_sampl
 				// if noisemode is 1, both taps are enabled
 				// if noisemode is 0, the lower tap, whitenoisetap2, is held at 0
 				// The != was a bit-XOR (^) before
-				if (((m_RNG & m_whitenoise_tap1)!=0) != (((m_RNG & m_whitenoise_tap2)!=(m_ncr_style_psg?m_whitenoise_tap2:0)) && in_noise_mode()))
-				{
+				if (((m_RNG & m_whitenoise_tap1) != 0) != ((static_cast<int32_t>(m_RNG & m_whitenoise_tap2) != (m_ncr_style_psg ? m_whitenoise_tap2 : 0)) && in_noise_mode())) {
 					m_RNG >>= 1;
 					m_RNG |= m_feedback_mask;
-				}
-				else
-				{
+				} else {
 					m_RNG >>= 1;
 				}
 				m_output[3] = m_RNG & 1;
